refactor: move mention classes to mention package
