+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/message/mention/MessageMention.kt
@@ -1,0 +1,9 @@
+package com.wire.kalium.logic.data.message.mention
+
+import com.wire.kalium.logic.data.user.UserId
+
+data class MessageMention(
+    val start: Int,
+    val length: Int,
+    val userId: UserId
+)
